+++ src/NavBar.php
@@ -4,9 +4,12 @@ declare(strict_types=1);
 
 namespace UIAwesome\Html\Component\Bootstrap5;
 
-use UIAwesome\Html\{Component\Bootstrap5\Cookbook\NavBar\Defaults, Core\Component\AbstractNavBar};
-
-use function array_merge;
+use PHPForge\Widget\Factory\SimpleFactory;
+use UIAwesome\Html\{
+    Component\Bootstrap5\Cookbook\NavBar\AlignRight,
+    Component\Bootstrap5\Cookbook\NavBar\Defaults,
+    Core\Component\AbstractNavBar
+};
 
 /**
  * A simple navbar component for displaying a navigation bar.
@@ -15,8 +18,26 @@ use function array_merge;
  */
 final class NavBar extends AbstractNavBar
 {
-    protected function loadDefaultDefinitions(): array
+    /**
+     * Define the navbar definition.
+     *
+     * @param string $definition The navbar definition. Available definitions: 'default', 'align-right'.
+     *
+     * @return self A new instance or clone of the current object with the applied definition.
+     */
+    public function definition(string $definition): self
     {
-        return array_merge(parent::loadDefaultDefinitions(), Defaults::definition());
+        $definition = match ($definition) {
+            'default' => Defaults::definition(),
+            'align-right' => AlignRight::definition(),
+            default => throw new \InvalidArgumentException(
+                sprintf(
+                    'Invalid definition: "%s". Available definitions: "default", "align-right".',
+                    $definition
+                )
+            ),
+        };
+
+        return SimpleFactory::configure($this, $definition);
     }
 }