+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 0.1.0 Under development
 
 - Enh #1: Initial commit (@terabytesoftw)
+- Enh #2: Add cookbook `NavBar` bootstrap5 `AlignRight` class (@terabytesoftw)